requirements.txt
beem
cryptography
pyyaml
requests
slackclient